requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.3.0
blinkpy==0.23.0
boto3==1.35.86
botocore==1.35.86
certifi==2024.12.14
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.5
jmespath==1.0.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
psutil==6.1.1
py-cpuinfo==9.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
ruff==0.8.4
s3transfer==0.10.4
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sortedcontainers==2.4.0
sympy==1.13.3
text-unidecode==1.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.53
ultralytics-thop==2.0.13
urllib3==1.26.20
yarl==1.18.3
zipp==3.21.0
